% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_intent_repos}
\alias{load_intent_repos}
\title{Load Repositories from DESCRIPTION}
\usage{
load_intent_repos(project, more_repos = NULL)
}
\arguments{
\item{project}{Path to the project directory.}

\item{more_repos}{Optional character vector of additional repositories.}
}
\value{
A named character vector of repositories.
}
\description{
Reads the repositories defined in \verb{Config/intent/repos/} and returns
them as a named character vector.
}
\keyword{internal}
