% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize an intent Project}
\usage{
init(path = ".", repos = NULL)
}
\arguments{
\item{path}{Character string. Path to the new project directory. Defaults to current working directory.}

\item{repos}{Character vector. Named repositories to use. Defaults to
\href{https://packagemanager.posit.co/cran/latest}{Posit Package Manager}
when neither this argument nor an existing DESCRIPTION file provides
repository configuration.}
}
\description{
Sets up a directory as an \code{intent} project by creating a \code{DESCRIPTION} file (if missing),
initializing a bare \code{renv} environment, setting the snapshot type to "explicit",
and configuring \code{.Rprofile} and \code{.Renviron}.
}
