% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intent-package.R
\docType{package}
\name{intent-package}
\alias{intent}
\alias{intent-package}
\title{intent: Fast, Unified Project Management for R}
\description{
A productivity-focused R package manager inspired by Python's 'uv'. Provides a single interface to manage project dependencies using the 'DESCRIPTION' file as intent and 'renv.lock' as state. Automates the complexity of keeping manifests, installed packages, and lockfiles in sync with simple commands like init(), add(), remove(), and sync().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SInginc/intent}
  \item Report bugs at \url{https://github.com/SInginc/intent/issues}
}

}
\author{
\strong{Maintainer}: Jack Liu \email{xiaochang.jack.liu@gmail.com}

}
\keyword{internal}
